Move RangeEditor.js to RangeEditor.jsx
This component uses jsx extensions.
