+++ media/js/src/editors/CostFunctionsUnitEditor.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import RangeEditor from '../form-components/RangeEditor.js';
+import RangeEditor from '../form-components/RangeEditor.jsx';
 import DefineRange from '../form-components/DefineRange.jsx';
 import { handleFormUpdate } from '../utils.js';
 import { getKatexEl } from '../katexUtils.jsx';
